+++ mobile/appcontainer/Toast.js
@@ -22,7 +22,7 @@ export const toast = hoistCmp.factory({
     model: uses(ToastModel),
 
     render({model}) {
-        const {icon, message, intent} = model,
+        const {icon, message, intent, actionButtonProps} = model,
             cls = `xh-toast xh-intent-${intent}`;
 
         return onsenToast({
@@ -32,6 +32,11 @@ export const toast = hoistCmp.factory({
             items: [
                 icon,
                 span(message),
+                button({
+                    omit: !actionButtonProps,
+                    minimal: true,
+                    ...actionButtonProps
+                }),
                 button({
                     icon: Icon.cross(),
                     minimal: true,